+++ packages/common/src/compiler/translator.ts
@@ -193,6 +193,17 @@ const convertDbtMetricToLightdashMetric = (
             sql = metric.sql;
         }
     }
+    if (metric.filters && metric.filters.length > 0) {
+        const filterSql = metric.filters
+            .map(
+                (filter) =>
+                    // eslint-disable-next-line no-useless-escape
+                    `(\$\{TABLE\}.${filter.field} ${filter.operator} ${filter.value})`,
+            )
+            .join(' AND ');
+        sql = `CASE WHEN ${filterSql} THEN ${sql} ELSE NULL END`;
+    }
+
     return {
         fieldType: FieldType.METRIC,
         type,