+++ packages/frontend/src/components/Explorer/VisualizationCard/ShowTotalsToggle.styles.ts
@@ -1,0 +1,5 @@
+import styled from 'styled-components';
+
+export const ShowTotalsToggleWrapper = styled.div`
+    margin-top: 5px;
+`;